Rename class and methods
