+++ lib/stack.rb
